Adding unittests for ePandda
These provide a testing framework for the API, the underlying databases
and the front-end at epandda.org.

At present this runs 8 tests across these services to confirm that they
are running, but could be extended to perform better checking of API
functionality and to be sure that the access points are returning
expected results.

The tests rely on the python unittest framework and selenium for
front-end testing. Selenium is using the chrome, which requires
both chrome and the chrome-driver library, both of which are easily
installable on Linux